move data source interfaces into folder
