+++ libs/repository/index.ts
@@ -1,5 +1,3 @@
-export * from './data-source-response.interface';
+export * from './data-source';
 export * from './detail-repository.interface';
 export * from './tree-repository.interface';
-export * from './tree-data-source.interface';
-export * from './data-source.interface';
